refactor(chat): extract provisionInputsMatch function to a separate module
This commit moves the `provisionInputsMatch` function from `shouldProvisionChatSession.ts` to a new file `provisionInputsMatch.ts`, improving modularity and code organization. The `useProvisionChatSession.ts` and related tests are updated to reflect this change, ensuring that the functionality remains intact while enhancing code clarity.
