+++ hooks/sessions/__tests__/shouldProvisionChatSession.test.ts
@@ -1,8 +1,5 @@
 import { describe, expect, it } from "vitest";
-import {
-  provisionInputsMatch,
-  shouldProvisionChatSession,
-} from "../shouldProvisionChatSession";
+import { shouldProvisionChatSession } from "../shouldProvisionChatSession";
 
 describe("shouldProvisionChatSession", () => {
   const base = {
@@ -68,27 +65,5 @@ describe("shouldProvisionChatSession", () => {
         isSuccess: true,
       }),
     ).toBe(true);
-  });
-});
-
-describe("provisionInputsMatch", () => {
-  it("matches when artistId and orgId equal lastVariables", () => {
-    expect(
-      provisionInputsMatch(
-        { artistId: "artist-a", orgId: "org-1" },
-        "artist-a",
-        "org-1",
-      ),
-    ).toBe(true);
-  });
-
-  it("does not match when artistId differs", () => {
-    expect(
-      provisionInputsMatch(
-        { artistId: "artist-a", orgId: undefined },
-        "artist-b",
-        undefined,
-      ),
-    ).toBe(false);
   });
 });
