+++ hooks/sessions/shouldProvisionChatSession.ts
@@ -1,4 +1,5 @@
 import type { ProvisionChatSessionInput } from "@/lib/sessions/provisionChatSession";
+import { provisionInputsMatch } from "@/hooks/sessions/provisionInputsMatch";
 
 interface ShouldProvisionChatSessionParams {
   enabled: boolean;
@@ -7,19 +8,6 @@ interface ShouldProvisionChatSessionParams {
   lastVariables: ProvisionChatSessionInput | undefined;
   isPending: boolean;
   isSuccess: boolean;
-}
-
-/** True when `lastVariables` matches the current `(artistId, orgId)` pair. */
-export function provisionInputsMatch(
-  lastVariables: ProvisionChatSessionInput | undefined,
-  artistId: string | undefined,
-  orgId: string | undefined,
-): boolean {
-  return (
-    lastVariables !== undefined &&
-    lastVariables.artistId === artistId &&
-    lastVariables.orgId === orgId
-  );
 }
 
 /**
